Maximum valued node in the path from node A to B in a bst.
Solution is to first Get X = LCA(A,B).
The find maxNodePath(X,A) and maxNodePath(X,B);
then find max of above two values;

Very interesting code to get LCA of nodes A and B.

Solution  LCS:

You  traverse down the tree if both A  and B belong to same subtree.
As soon as they belong to different subtrees or one of them is equal to
curret node we are visiting, then we are the LCA. This O(logn) time
solution to get LCS with  O(1) space.