Convert  BST to double linked list in place:
This turned out  to  be an easy problem.

Once you are at  a  node, traverse the tree inorderly so  that we visit
nodes  in ascending order, then once you are  back from leftSubtree and
your head is  still empty, that  means this  node  is head and tail as
of NOW.  If Head is not NULL, then add this node to the tail, with two
links   tail->root   root->tail. and  recurse  to the right after
updating  root as tail.

There  are 4 algorithms  in GFG to convert BST to DLL.
Let's see them one by one.